Fix norminette problems in parcer files
